Move multi upstream/downstream induction functions
